requirements.txt
amadeus==10.1.0
asgiref==3.8.1
async-timeout==4.0.3
boto3==1.34.153
botocore==1.34.153
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.0
Django==5.0.7
django-admin-logs==1.2.0
django-cors-headers==4.4.0
django-oauth-toolkit==2.4.0
django-redis==5.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docker==7.1.0
drf-yasg==1.21.7
exceptiongroup==1.2.2
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
jmespath==1.0.1
jwcrypto==1.5.6
mock==5.1.0
oauthlib==3.2.2
packaging==24.1
pluggy==1.5.0
pycparser==2.22
PyJWT==2.8.0
pytest==8.3.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
